#63 - refactor: 채팅방 Service 단일 책임 원칙 적용
